Clean harness utility layout
