+++ dotenv/tests/integration/dotenv.rs
@@ -4,7 +4,7 @@ use crate::util::*;
 fn dotenv_ok_default_env() {
     test_in_default_env(|| {
         dotenvy::dotenv().ok();
-        assert_env_var(TEST_EXISTING_KEY, TEST_EXISTING_VALUE);
-        assert_env_var(TEST_KEY, TEST_VALUE);
+        assert_env_var(DEFAULT_EXISTING_KEY, DEFAULT_EXISTING_VALUE);
+        assert_env_var(DEFAULT_TEST_KEY, DEFAULT_TEST_VALUE);
     });
 }
