+++ dotenv/tests/integration/util/tests/default_env.rs
@@ -3,8 +3,8 @@ use super::*;
 #[test]
 fn vars_state() {
     test_in_default_env(|| {
-        assert_env_var(TEST_EXISTING_KEY, TEST_EXISTING_VALUE);
-        assert_env_var_unset(TEST_KEY);
+        assert_env_var(DEFAULT_EXISTING_KEY, DEFAULT_EXISTING_VALUE);
+        assert_env_var_unset(DEFAULT_TEST_KEY);
     });
 }
 
@@ -19,7 +19,7 @@ fn envfile_loaded_vars_state() {
     test_in_default_env(|| {
         dotenv_wrap().expect(DOTENV_EXPECT);
         // dotenv() does not override existing var
-        assert_env_var(TEST_EXISTING_KEY, TEST_EXISTING_VALUE);
-        assert_env_var(TEST_KEY, TEST_VALUE);
+        assert_env_var(DEFAULT_EXISTING_KEY, DEFAULT_EXISTING_VALUE);
+        assert_env_var(DEFAULT_TEST_KEY, DEFAULT_TEST_VALUE);
     });
 }
