+++ dotenv/tests/integration/util/tests/envfile_builder.rs
@@ -15,27 +15,27 @@ fn default_builds_empty() {
 #[test]
 fn add_key_empty_value() {
     let mut efb = EnvFileBuilder::new();
-    efb.add_key_value(TEST_KEY, "");
-    let expected = format!("{}=\n", TEST_KEY);
+    efb.add_key_value(DEFAULT_TEST_KEY, "");
+    let expected = format!("{}=\n", DEFAULT_TEST_KEY);
     assert_contents_str(efb, &expected);
 }
 
 #[test]
 fn add_key_value() {
     let mut efb = EnvFileBuilder::new();
-    efb.add_key_value(TEST_KEY, TEST_VALUE);
-    let expected = format!("{}={}\n", TEST_KEY, TEST_VALUE);
+    efb.add_key_value(DEFAULT_TEST_KEY, DEFAULT_TEST_VALUE);
+    let expected = format!("{}={}\n", DEFAULT_TEST_KEY, DEFAULT_TEST_VALUE);
     assert_contents_str(efb, &expected);
 }
 
 #[test]
 fn add_multiple_key_values() {
     let mut efb = EnvFileBuilder::new();
-    efb.add_key_value(TEST_KEY, TEST_VALUE);
-    efb.add_key_value(TEST_EXISTING_KEY, TEST_OVERRIDING_VALUE);
+    efb.add_key_value(DEFAULT_TEST_KEY, DEFAULT_TEST_VALUE);
+    efb.add_key_value(DEFAULT_EXISTING_KEY, TEST_OVERRIDING_VALUE);
     let expected = expected_envfile(&[
-        (TEST_KEY, TEST_VALUE),
-        (TEST_EXISTING_KEY, TEST_OVERRIDING_VALUE),
+        (DEFAULT_TEST_KEY, DEFAULT_TEST_VALUE),
+        (DEFAULT_EXISTING_KEY, TEST_OVERRIDING_VALUE),
     ]);
     assert_contents_str(efb, &expected);
 }
